pyproject.toml
[project]
name = "sunnieclassroomadventures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "onnxscript>=0.6.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
